pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "taproot-rca"
version = "0.1.0"
description = "AI-powered schema drift detection and self-healing for data engineers, using local LLMs via Ollama"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Taproot-RCA Contributors" },
]
keywords = ["schema", "drift", "data-engineering", "ollama", "self-healing", "rca"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Database",
    "Topic :: Software Development :: Quality Assurance",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
]

dependencies = [
    "typer[all]>=0.9.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "httpx>=0.25.0",
    "rich>=13.0",
]

[project.optional-dependencies]
postgres = ["psycopg2-binary>=2.9"]
mysql = ["pymysql>=1.1"]
snowflake = ["snowflake-connector-python>=3.0"]
all-connectors = [
    "psycopg2-binary>=2.9",
    "pymysql>=1.1",
    "snowflake-connector-python>=3.0",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.1.0",
]

[project.scripts]
taproot = "taproot_rca.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/taproot_rca"]

[tool.ruff]
line-length = 100
target-version = "py310"
